Avoid instance_eval by patching all the request helpers
Creates a `RequestHelpers` module in a distinct `Spec::Support::Api`
namespace (and moves the `ApiHelper` in there for consistency) with
patched versions of the request helpers provided by Rails.
